refactor: remove common by moving files into application
