+++ src/apps/backend/modules/account/internal/store/account_model.py
@@ -3,7 +3,7 @@ from typing import Optional
 from bson import ObjectId
 from dataclasses import dataclass
 from modules.account.types import PhoneNumber
-from modules.common.base_model import BaseModel
+from modules.application.base_model import BaseModel
 
 
 @dataclass
